move classes to domain package
