+++ samples/java-spring-customer-registry-views-quickstart/src/main/java/customer/domain/Address.java
@@ -1,4 +1,4 @@
-package customer.api;
+package customer.domain;
 
 public record Address(String street, String city) {
 }
