+++ samples/java-spring-customer-registry-views-quickstart/src/main/java/customer/domain/Customer.java
@@ -1,4 +1,4 @@
-package customer.api;
+package customer.domain;
 
 // tag::class[]
 public record Customer(String customerId, String email, String name, Address address) {
