+++ samples/java-spring-customer-registry-views-quickstart/src/main/java/customer/view/CustomerByNameView.java
@@ -1,7 +1,7 @@
 package customer.view;
 
 // tag::class[]
-import customer.api.Customer;
+import customer.domain.Customer;
 import customer.api.CustomerEntity;
 import kalix.javasdk.view.View;
 import kalix.javasdk.annotations.Query;
